% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastQ_to_bigwig_to_peaks.R
\name{fastQ_to_bigwig_to_peaks}
\alias{fastQ_to_bigwig_to_peaks}
\title{ChIPseq Pipeline (fastq to bw to peaks)}
\usage{
fastQ_to_bigwig_to_peaks(fastq_path, mapq = 15, buildIndex = FALSE,
  index_basename, sample_sheet_path, bw = 300, q = 0.05, genome,
  blacklist_path, onlyCallPeaks = FALSE, bam_files_callPeaks)
}
\arguments{
\item{fastq_path}{path to the folder wehre the fastqs are}

\item{mapq}{mapping quality score (mapq) used for filtering BAMs. The default is mapq = 15, which is what ChIPQC uses as its cut off}

\item{buildIndex}{if TRUE this will build the index for the Rsubread alignment, only need to do this once.}

\item{index_basename}{the name of the index file for alignment, assuming it isn't built within the pipeline}

\item{sample_sheet_path}{Path to the sample sheet for peak calling. If this argument is not included, no peaks will be called.}

\item{bw}{MACS2 argument if calling peaks, default is 300}

\item{q}{MACS2 argument if calling peaks, default is 0.05}

\item{genome}{gemone used for alignments, etc}

\item{blacklist_path}{path to blacklist bed file (include this as data once you know how to do that?)}

\item{onlyCallPeaks}{if you only want to call peaks and not do alignments, set this to true. You need the sample_sheet_path to have an argument as well}
}
\value{
many things
}
\description{
Dougs pipeline that takes a fastq, aligns, creates a bigwig, then calls peaks if desired, with a lot of fun QC along the way
}
\details{
NA
}
\examples{
message("coming soon!")
}
